PillMeIn-OCR/requirements.txt
Flask
google-cloud-vision
openai
flasgger